Data.Vinyl

Documentation

module Data.Vinyl.Core

class RecMapMethod c f ts where

Apply a typeclass method to each field of a Rec where the class constrains the index of the field, but not its interpretation functor.

Methods

rmapMethod :: (forall a. c (PayloadType f a) => f a -> g a) -> Rec f ts -> Rec g ts

class RecPointed c f ts where

Generate a record from fields derived from type class instances.

Methods

rpointMethod :: (forall a. c (f a) => f a) -> Rec f ts

rmapMethodF :: forall c f ts. (Functor f, FieldPayload f ~ 'FieldId, RecMapMethod c f ts) => (forall a. c a => a -> a) -> Rec f ts -> Rec f ts

Apply a typeclass method to each field of a Rec f ts using the Functor instance for f to lift the function into the functor. This is a commonly-used specialization of rmapMethod composed with fmap.

mapFields :: forall c ts. RecMapMethod c ElField ts => (forall a. c a => a -> a) -> FieldRec ts -> FieldRec ts

Apply a typeclass method to each field of a FieldRec. This is a specialization of rmapMethod.

rtraverseInMethod :: forall c h f g rs. (RMap rs, RPureConstrained c rs, RApply rs) => (forall a. c a => f a -> g (ApplyToField h a)) -> Rec f rs -> Rec g (MapTyCon h rs)

Like rtraverseIn, but the function between functors may be constrained.

rsequenceInFields :: forall f rs. (Functor f, AllFields rs, RMap rs) => Rec (f :. ElField) rs -> Rec ElField (MapTyCon f rs)

Push an outer layer of interpretation functor into each named field.

data ARec (f :: k -> *) (ts :: [k])

An array-backed extensible record with constant-time field access.

toARec :: forall f ts. NatToInt (RLength ts) => Rec f ts -> ARec f ts

Convert a Rec into an ARec for constant-time field access.

fromARec :: forall f ts. (RecApplicative ts, RPureConstrained (IndexableField ts) ts) => ARec f ts -> Rec f ts

Convert an ARec into a Rec.

module Data.Vinyl.Derived

record :: TupleRec f t => t -> UncurriedRec (TupleToRecArgs f t)

xrecX :: TupleXRec f t => ListToHKDTuple f t -> XRec f t

Build an XRec from a tuple.

xrecTuple :: TupleXRec f t => XRec f t -> ListToHKDTuple f t

Convert an XRec to a tuple. Useful for pattern matching on an entire record.

ruple :: (IsoXRec f ts, TupleXRec f ts) => Rec f ts -> ListToHKDTuple f ts

Convert a Rec to a tuple going through HKD to reduce syntactic noise. Useful for pattern matching on an entire Rec.

xrec :: (IsoXRec f t, TupleXRec f t) => ListToHKDTuple f t -> Rec f t

Build a Rec from a tuple passing through XRec. This admits the most concise syntax for building a Rec. For example, xrec ("joe", 23) :: Rec Identity '[String, Int].

fieldRec :: TupleRec ElField t => t -> UncurriedRec (TupleToRecArgs ElField t)

Build a FieldRec from a tuple of ElField values.

data ElField (field :: (Symbol, Type)) where

A value with a phantom Symbol label. It is not a Haskell Functor, but it is used in many of the same places a Functor is used in vinyl.

Constructors

Field :: KnownSymbol s => !t -> ElField '(s, t)

module Data.Vinyl.Lens

data SRec f ts

A simpler type for SRec2 whose RecElem and RecSubset instances are specialized to the ElField functor.

toSRec :: Storable (Rec f ts) => Rec f ts -> SRec f ts

Create an SRec from a Rec. This should offer very fast field access, but note that its lens API (via RecElem and RecSubset) is restricted to the ElField functor.

fromSRec :: Storable (Rec f ts) => SRec f ts -> Rec f ts

Create a Rec from an SRec.

class IsoXRec f ts where

Conversion between XRec and Rec. It is convenient to build and consume XRec values to reduce syntactic noise, but Rec has a richer API that is difficult to build around the HKD type family.

Methods

fromXRec :: XRec f ts -> Rec f ts

toXRec :: Rec f ts -> XRec f ts

class XRApply f g rs where

Like rapply: record of components f r -> g r may be applied to a record of f to get a record of g.

Methods

xrapply :: XRec (Lift (->) f g) rs -> XRec f rs -> XRec g rs

class XRMap f g rs

The implementation of xrmap is broken into a type class to permit unrolling of the recursion across a record. The function mapped across the vector hides the HKD type family under a newtype constructor to help the type checker.

Minimal complete definition

xrmapAux

type XRec f = Rec (XData f)

pattern XRNil :: XRec f '[]

pattern (::&) :: HKD f r -> XRec f rs -> XRec f (r ': rs) infixr 7

rmapX :: forall f g rs. (XRMap f g rs, IsoXRec f rs, IsoXRec g rs) => (forall a. HKD f a -> HKD g a) -> Rec f rs -> Rec g rs

Like rmap, but the supplied function is written against the HKD-simplified types. This is xrmap sandwiched in between fromXRec and toXRec.

xrmap :: forall f g rs. XRMap f g rs => (forall a. HKD f a -> HKD g a) -> XRec f rs -> XRec g rs

This is rmap for XRec. We apply a natural transformation between interpretation functors to transport a record value between interpretations.