Data.Vinyl.ARec.Internal

Description

Constant-time field accessors for extensible records. The trade-off is the usual lists vs arrays one: it is fast to add an element to the head of a list, but element access is linear time; array access time is uniform, but extending the array is more slower.

Tradeoffs:

No sharing of the spine (i.e. when you change elements in the front of the record the tail can't be re-used)

ARec requires (4 + n) words + size of the fields

1 for the ARec constructor

1 for the pointer to the SmallArray#

The SmallArray# has 2 words as header (1 for GC, 1 for number of elements)

1 pointer per element to the actual data

Rec requires (2n) words + size of Fields

1 word per (:&) constructor

1 word for the pointer to the element

Documentation

newtype ARec (f :: k -> *) (ts :: [k])

An array-backed extensible record with constant-time field access.

Constructors

ARec SmallArray

class ToARec (us :: [k])

Minimal complete definition

aRecValues

class NatToInt (RIndex t ts) => IndexableField ts t

Defines a constraint that lets us index into an ARec in order to produce a Rec using fromARec.

arec :: forall k (us :: [k]) f. NatToInt (RLength us) => ARecBuilder f us -> ARec f us

Turn an ARecBuilder into an ARec

See ARecBuilder

newtype ARecBuilder f us

An efficient builder for ARec values

Use the pseudo-constructors arcons and arnil to construct an ARecBuilder and then turn it into an ARec with arec

Example: (requires -XOverloadedLabels and )

user :: ARec ElField '[ "name"   ::: String
                      , "age"    ::: Int
                      , "active" ::: Bool]
user = arec (  #name   =: "Peter"
            `arcons` #age    =: 4
            `arcons` #active =: True
            `arcons` arnil
            )

Constructors

ARecBuilder (forall s. Int -> SmallMutableArray s -> ST s ())

arcons :: f u -> ARecBuilder f us -> ARecBuilder f (u ': us) infixr 1

Pseudo-constructor for an ARecBuilder

Cons a field to an ARec under construction

See ARecBuilder

arnil :: ARecBuilder f '[]

Pseudo-constructor for ARecBuilder

Build an ARec without fields

See ARecBuilder

toARec :: forall f ts. (NatToInt (RLength ts), ToARec ts) => Rec f ts -> ARec f ts

Convert a Rec into an ARec for constant-time field access.

fromARec :: forall f ts. (RecApplicative ts, RPureConstrained (IndexableField ts) ts) => ARec f ts -> Rec f ts

Convert an ARec into a Rec.

aget :: forall t f ts. NatToInt (RIndex t ts) => ARec f ts -> f t

Get a field from an ARec.

unsafeAput :: forall t t' f ts ts'. NatToInt (RIndex t ts) => f t' -> ARec f ts -> ARec f ts'

Set a field in an ARec.

unsafeAlens :: forall f g t t' ts ts'. (Functor g, NatToInt (RIndex t ts)) => (f t -> g (f t')) -> ARec f ts -> g (ARec f ts')

Define a lens for a field of an ARec.

arecGetSubset :: forall rs ss f. (IndexWitnesses (RImage rs ss), NatToInt (RLength rs)) => ARec f ss -> ARec f rs

Get a subset of a record's fields.

arecSetSubset :: forall rs ss f. IndexWitnesses (RImage rs ss) => ARec f ss -> ARec f rs -> ARec f ss

Set a subset of a larger record's fields to all of the fields of a smaller record.

arecRepsMatchCoercion :: AllRepsMatch f xs g ys => Coercion (ARec f xs) (ARec g ys)

Given that xs and ys have the same length, and mapping f over xs and g over ys produces lists whose elements are pairwise Coercible, ARec f xs and ARec g ys are Coercible.

arecConsMatchCoercion :: (forall (x :: k). Coercible (f x) (g x)) => Coercion (ARec f xs) (ARec g xs)

Given that forall x. Coercible (f x) (g x), produce a coercion from ARec f xs to ARec g xs. While the constraint looks a lot like Coercible f g, it is actually weaker.