Data.Vinyl.ARec.Internal.SmallArray

Description

Helper functions for SmallArray#

This module exposes _unsafe_ functions to work with SmallArrays. That means that specifically neither index bounds nor element types are checked So this functionality should only be used in a context that enforces them by some other means, e.g. ARec's type index

Documentation

data SmallArray

Constructors

SmallArray !(SmallArray# Any)

data SmallMutableArray s

Constructors

SmallMutableArray !(SmallMutableArray# s Any)

indexSmallArray :: SmallArray -> Int -> a

withNewSmallArray :: Int -> (SmallMutableArray s -> ST s ()) -> ST s SmallArray

writeSmallArray :: SmallMutableArray s -> Int -> a -> ST s ()

withThawedSmallArray :: SmallArray -> (SmallMutableArray s -> ST s ()) -> ST s SmallArray