Data.Vinyl.Core

Description

Core vinyl definitions. The Rec data type is defined here, but also of interest are definitions commonly used functions like rmap, rapply, and rtraverse.

The definitions in this module are written in terms of type classes so that the definitions may be specialized to each record type at which they are used. This usually helps with runtime performance, but can slow down compilation time. If you are experiencing poor compile times, you may wish to try the semantically equivalent definitions in the Data.Vinyl.Recursive module: they should produce the same results given the same inputs as functions defined in this module, but they will not be specialized to your record type. Instead, they treat the record as a list of fields, so will have performance linear in the size of the record.

Documentation

data Rec :: (u -> *) -> [u] -> * where

A record is parameterized by a universe u, an interpretation f and a list of rows rs. The labels or indices of the record are given by inhabitants of the kind u; the type of values at any label r :: u is given by its interpretation f r :: *.

Constructors

RNil :: Rec f '[]

(:&) :: !(f r) -> !(Rec f rs) -> Rec f (r ': rs) infixr 7

rappend :: Rec f as -> Rec f bs -> Rec f (as ++ bs)

Two records may be pasted together.

(<+>) :: Rec f as -> Rec f bs -> Rec f (as ++ bs) infixr 5

A shorthand for rappend.

rcombine :: (RMap rs, RApply rs) => (forall a. m a -> m a -> m a) -> (forall a. f a -> m a) -> (forall a. m a -> g a) -> Rec f rs -> Rec f rs -> Rec g rs

Combine two records by combining their fields using the given function. The first argument is a binary operation for combining two values (e.g. (<>)), the second argument takes a record field into the type equipped with the desired operation, the third argument takes the combined value back to a result type.

class RMap rs where

Rec _ rs with labels in kind u gives rise to a functor Hask^u -> Hask; that is, a natural transformation between two interpretation functors f,g may be used to transport a value from Rec f rs to Rec g rs.

Methods

rmap :: (forall x. f x -> g x) -> Rec f rs -> Rec g rs

(<<$>>) :: RMap rs => (forall x. f x -> g x) -> Rec f rs -> Rec g rs infixl 8

A shorthand for rmap.

(<<&>>) :: RMap rs => Rec f rs -> (forall x. f x -> g x) -> Rec g rs

An inverted shorthand for rmap.

class RApply rs where

A record of components f r -> g r may be applied to a record of f to get a record of g.

Methods

rapply :: Rec (Lift (->) f g) rs -> Rec f rs -> Rec g rs

(<<*>>) :: RApply rs => Rec (Lift (->) f g) rs -> Rec f rs -> Rec g rs infixl 8

A shorthand for rapply.

class RecApplicative rs where

Given a section of some functor, records in that functor of any size are inhabited.

Methods

rpure :: (forall x. f x) -> Rec f rs

rtraverse :: Applicative h => (forall x. f x -> h (g x)) -> Rec f rs -> h (Rec g rs)

A record may be traversed with respect to its interpretation functor. This can be used to yank (some or all) effects from the fields of the record to the outside of the record.

rtraverseIn :: forall h f g rs. (forall a. f a -> g (ApplyToField h a)) -> Rec f rs -> Rec g (MapTyCon h rs)

While rtraverse pulls the interpretation functor out of the record, rtraverseIn pushes the interpretation functor in to each field type. This is particularly useful when you wish to discharge that interpretation on a per-field basis. For instance, rather than a Rec IO '[a,b], you may wish to have a Rec Identity '[IO a, IO b] so that you can evaluate a single field to obtain a value of type Rec Identity '[a, IO b].

rsequenceIn :: forall f g (rs :: [Type]). (Traversable f, Applicative g) => Rec (f :. g) rs -> Rec g (MapTyCon f rs)

Push an outer layer of interpretation functor into each field.

rzipWith :: (RMap xs, RApply xs) => (forall x. f x -> g x -> h x) -> Rec f xs -> Rec g xs -> Rec h xs

Given a natural transformation from the product of f and g to h, we have a natural transformation from the product of Rec f and Rec g to Rec h. You can also think about this operation as zipping two records with the same element types but different interpretations.

class RFoldMap rs where

Map each element of a record to a monoid and combine the results.

Methods

rfoldMapAux :: Monoid m => (forall x. f x -> m) -> m -> Rec f rs -> m

rfoldMap :: forall rs m f. (Monoid m, RFoldMap rs) => (forall x. f x -> m) -> Rec f rs -> m

class RecordToList rs where

A record with uniform fields may be turned into a list.

Methods

recordToList :: Rec (Const a) rs -> [a]

data Dict c a where

Wrap up a value with a capability given by its type

Constructors

Dict :: c a => a -> Dict c a

class ReifyConstraint c f rs where

Sometimes we may know something for all fields of a record, but when you expect to be able to each of the fields, you are then out of luck. Surely given ∀x:u.φ(x) we should be able to recover x:u ⊢ φ(x)! Sadly, the constraint solver is not quite smart enough to realize this and we must make it patently obvious by reifying the constraint pointwise with proof.

Methods

reifyConstraint :: Rec f rs -> Rec (Dict c :. f) rs

class RPureConstrained c ts where

Build a record whose elements are derived solely from a constraint satisfied by each.

Methods

rpureConstrained :: (forall a. c a => f a) -> Rec f ts

data DictOnly (c :: k -> Constraint) a where

Capture a type class instance dictionary. See getDict for a way to obtain a DictOnly value from an RPureConstrained constraint.

Constructors

DictOnly :: forall c a. c a => DictOnly c a

withPairedDict :: (c a => f a -> r) -> Product (DictOnly c) f a -> r

A useful technique is to use 'rmap (Pair (DictOnly MyClass))' on a Rec to pair each field with a type class dictionary for MyClass@. This helper can then be used to eliminate the original.

class RPureConstraints cs ts where

Build a record whose elements are derived solely from a list of constraint constructors satisfied by each.

Methods

rpureConstraints :: (forall a. AllSatisfied cs a => f a) -> Rec f ts

type family Head xs where ...

Equations

Head (x ': _) = x

type family Tail xs where ...

Equations

Tail (_ ': xs) = xs

type family AllRepsMatch_ (f :: j -> *) (xs :: [j]) (g :: k -> *) (ys :: [k]) :: Constraint where ...

Equations

AllRepsMatch_ f (x ': xs) g ys = (ys ~ (Head ys ': Tail ys), Coercible (f x) (g (Head ys)), AllRepsMatch_ f xs g (Tail ys))

AllRepsMatch_ _ '[] _ ys = ys ~ '[]

type AllRepsMatch f xs g ys = (AllRepsMatch_ f xs g ys, AllRepsMatch_ g ys f xs)

AllRepsMatch f xs g ys means that xs and ys have the same lengths, and that mapping f over xs and g over ys produces lists whose corresponding elements are Coercible with each other. For example, the following hold:

AllRepsMatch Proxy '[1,2,3] Proxy '[4,5,6] AllRepsMatch Sum '[Int,Word] Identity '[Min Int, Max Word]

repsMatchCoercion :: AllRepsMatch f xs g ys => Coercion (Rec f xs) (Rec g ys)

Given that for each element x in the list xs,

consMatchCoercion :: (forall (x :: k). Coercible (f x) (g x)) => Coercion (Rec f xs) (Rec g xs)