Data.Vinyl.Derived

Description

Commonly used Rec instantiations.

Documentation

type (:::) a b = '(a, b)

Alias for Field spec

type FieldRec = Rec ElField

A record of named fields.

type AFieldRec ts = ARec ElField ts

An ARec of named fields to provide constant-time field access.

type HList = Rec Identity

Heterogeneous list whose elements are evaluated during list construction.

type LazyHList = Rec Thunk

Heterogeneous list whose elements are left as-is during list construction (cf. HList).

getField :: ElField '(s, t) -> t

Get the data payload of an ElField.

getLabel :: forall s t. ElField '(s, t) -> String

Get the label name of an ElField.

fieldMap :: (a -> b) -> ElField '(s, a) -> ElField '(s, b)

ElField is isomorphic to a functor something like Compose ElField ('(,) s).

traverseField :: (KnownSymbol s, Functor f) => (a -> b) -> f (ElField '(s, a)) -> ElField '(s, f b)

Something in the spirit of traverse for ElField whose kind fights the standard library.

rfield :: Functor f => (a -> f b) -> ElField '(s, a) -> f (ElField '(s, b))

Lens for an ElField's data payload.

(=:) :: KnownSymbol l => Label (l :: Symbol) -> (v :: *) -> ElField (l ::: v) infix 8

Operator for creating an ElField. With the -XOverloadedLabels extension, this permits usage such as, #foo =: 23 to produce a value of type ElField ("foo" ::: Int).

rgetf :: forall l f v record us. (HasField record l us us v v, RecElemFCtx record f) => Label l -> record f us -> f (l ::: v)

Get a named field from a record.

rvalf :: (HasField record l us us v v, RecElemFCtx record ElField) => Label l -> record ElField us -> v

Get the value associated with a named field from a record.

rputf' :: forall l v v' record us us'. (HasField record l us us' v v', KnownSymbol l, RecElemFCtx record ElField) => Label l -> v' -> record ElField us -> record ElField us'

Set a named field. rputf' #foo 23 sets the field named #foo to 23.

rputf :: forall l v record us. (HasField record l us us v v, KnownSymbol l, RecElemFCtx record ElField) => Label l -> v -> record ElField us -> record ElField us

Set a named field without changing its type. rputf #foo 23 sets the field named #foo to 23.

rlensfL' :: forall l v v' record g f us us'. (Functor g, HasField record l us us' v v', RecElemFCtx record f) => Label l -> (f (l ::: v) -> g (f (l ::: v'))) -> record f us -> g (record f us')

A lens into a Rec identified by a Label.

rlensfL :: forall l v record g f us. (Functor g, HasField record l us us v v, RecElemFCtx record f) => Label l -> (f (l ::: v) -> g (f (l ::: v))) -> record f us -> g (record f us)

A type-preserving lens into a Rec identified by a Label.

rlensf' :: forall l v v' record g us us'. (Functor g, HasField record l us us' v v', RecElemFCtx record ElField) => Label l -> (v -> g v') -> record ElField us -> g (record ElField us')

A lens into the payload value of a Rec field identified by a Label.

rlensf :: forall l v record g us. (Functor g, HasField record l us us v v, RecElemFCtx record ElField) => Label l -> (v -> g v) -> record ElField us -> g (record ElField us)

A type-preserving lens into the payload value of a Rec field identified by a Label.

(=:=) :: KnownSymbol s => Label (s :: Symbol) -> a -> FieldRec '['(s, a)]

Shorthand for a FieldRec with a single field.

data SField (field :: k)

A proxy for field types.

Constructors

SField

type family FieldType l fs where ...

Equations

FieldType l '[] = TypeError (('Text "Cannot find label " :<>: 'ShowType l) :<>: 'Text " in fields")

FieldType l ((l ::: v) ': fs) = v

FieldType l ((l' ::: v') ': fs) = FieldType l fs

type HasField record l fs fs' v v' = (RecElem record (l ::: v) (l ::: v') fs fs' (RIndex (l ::: v) fs), FieldType l fs ~ v, FieldType l fs' ~ v')

Constraint that a label is associated with a particular type in a record.

data Label (a :: Symbol)

Proxy for label type

Constructors

Label

class (KnownSymbol (Fst a), a ~ '(Fst a, Snd a)) => KnownField a

Defines a constraint that lets us extract the label from an ElField. Used in rmapf and rpuref.

type AllFields fs = (RPureConstrained KnownField fs, RecApplicative fs, RApply fs)

Shorthand for working with records of fields as in rmapf and rpuref.

rmapf :: AllFields fs => (forall a. KnownField a => f a -> g a) -> Rec f fs -> Rec g fs

Map a function between functors across a Rec taking advantage of knowledge that each element is an ElField.

type family Unlabeled ts where ...

Remove the first component (e.g. the label) from a type-level list of pairs.

Equations

Unlabeled '[] = '[]

Unlabeled ('(s, x) ': xs) = x ': Unlabeled xs

class StripFieldNames ts where

Facilities for removing and replacing the type-level label, or column name, part of a record.

Methods

stripNames :: Rec ElField ts -> Rec Identity (Unlabeled ts)

stripNames' :: Functor f => Rec (f :. ElField) ts -> Rec f (Unlabeled ts)

withNames :: Rec Identity (Unlabeled ts) -> Rec ElField ts

withNames' :: Functor f => Rec f (Unlabeled ts) -> Rec (f :. ElField) ts

rpuref :: AllFields fs => (forall a. KnownField a => f a) -> Rec f fs

Construct a Rec with ElField elements.

(<<$$>>) :: AllFields fs => (forall a. KnownField a => f a -> g a) -> Rec f fs -> Rec g fs

Operator synonym for rmapf.

rlabels :: AllFields fs => Rec (Const String) fs

Produce a Rec of the labels of a Rec of ElFields.

Specializations for working with an ARec of named fields.