Data.Vinyl.Lens

Description

Lenses into record fields.

Documentation

class (i ~ RIndex r rs, NatToInt i) => RecElem (record :: (k -> *) -> [k] -> *) (r :: k) (r' :: k) (rs :: [k]) (rs' :: [k]) (i :: Nat) | r r' rs i -> rs' where

The presence of a field in a record is witnessed by a lens into its value. The fifth parameter to RecElem, i, is there to help the constraint solver realize that this is a decidable predicate with respect to the judgemental equality in k.

Associated Types

type RecElemFCtx record (f :: k -> *) :: Constraint

An opportunity for instances to generate constraints based on the functor parameter of records passed to class methods.

type RecElemFCtx record f = ()

Methods

rlensC :: (Functor g, RecElemFCtx record f) => (f r -> g (f r')) -> record f rs -> g (record f rs')

We can get a lens for getting and setting the value of a field which is in a record. As a convenience, we take a proxy argument to fix the particular field being viewed. These lenses are compatible with the lens library. Morally:

rlensC :: Lens' (Rec f rs) (Rec f rs') (f r) (f r')

rgetC :: (RecElemFCtx record f, r ~ r') => record f rs -> f r

For Vinyl users who are not using the lens package, we provide a getter.

rputC :: RecElemFCtx record f => f r' -> record f rs -> record f rs'

For Vinyl users who are not using the lens package, we also provide a setter. In general, it will be unambiguous what field is being written to, and so we do not take a proxy argument here.

rget :: forall r rs f record. (RecElem record r r rs rs (RIndex r rs), RecElemFCtx record f) => record f rs -> f r

rput :: forall k (r :: k) rs record f. (RecElem record r r rs rs (RIndex r rs), RecElemFCtx record f) => f r -> record f rs -> record f rs

Type-preserving field setter. This type is simpler to work with than that of rput'.

rput' :: forall k (r :: k) (r' :: k) (rs :: [k]) (rs' :: [k]) record f. (RecElem record r r' rs rs' (RIndex r rs), RecElemFCtx record f) => f r' -> record f rs -> record f rs'

The type-changing field setter rputC with the type arguments re-ordered for more convenient usage with TypeApplications.

rlens :: forall r record rs f g. (RecElem record r r rs rs (RIndex r rs), RecElemFCtx record f, Functor g) => (f r -> g (f r)) -> record f rs -> g (record f rs)

Type-preserving field lens. This type is simpler to work with than that of rlens'.

rlens' :: forall r r' record rs rs' f g. (RecElem record r r' rs rs' (RIndex r rs), RecElemFCtx record f, Functor g) => (f r -> g (f r')) -> record f rs -> g (record f rs')

Type-changing field lens rlensC with the type arguments re-ordered for more convenient usage with TypeApplications.

type RElem x rs = RecElem Rec x x rs rs

RecElem for classic vinyl Rec types.

class is ~ RImage rs ss => RecSubset record rs ss is where

If one field set is a subset another, then a lens of from the latter's record to the former's is evident. That is, we can either cast a larger record to a smaller one, or we may replace the values in a slice of a record.

Minimal complete definition

rsubsetC

Associated Types

type RecSubsetFCtx record (f :: k -> *) :: Constraint

An opportunity for instances to generate constraints based on the functor parameter of records passed to class methods.

type RecSubsetFCtx record f = ()

Methods

rsubsetC :: (Functor g, RecSubsetFCtx record f) => (record f rs -> g (record f rs)) -> record f ss -> g (record f ss)

This is a lens into a slice of the larger record. Morally, we have:

rsubset :: Lens' (Rec f ss) (Rec f rs)

rcastC :: RecSubsetFCtx record f => record f ss -> record f rs

The getter of the rsubset lens is rcast, which takes a larger record to a smaller one by forgetting fields.

rreplaceC :: RecSubsetFCtx record f => record f rs -> record f ss -> record f ss

The setter of the rsubset lens is rreplace, which allows a slice of a record to be replaced with different values.

rsubset :: forall k rs ss f g record is. (RecSubset record (rs :: [k]) (ss :: [k]) is, Functor g, RecSubsetFCtx record f) => (record f rs -> g (record f rs)) -> record f ss -> g (record f ss)

A lens into a slice of the larger record. This is rsubsetC with the type arguments reordered for more convenient usage with TypeApplications.

rcast :: forall rs ss f record is. (RecSubset record rs ss is, RecSubsetFCtx record f) => record f ss -> record f rs

Takes a larger record to a smaller one by forgetting fields. This is rcastC with the type arguments reordered for more convenient usage with TypeApplications.

rreplace :: forall rs ss f record is. (RecSubset record rs ss is, RecSubsetFCtx record f) => record f rs -> record f ss -> record f ss

Allows a slice of a record to be replaced with different values. This is rreplaceC with the type arguments reordered for more convenient usage with TypeApplications.

rdowncast :: (RecApplicative ss, RMap rs, rs ⊆ ss) => Rec f rs -> Rec (Maybe :. f) ss

Takes a smaller record to a larger one, a downcast, by layering a Maybe interpretation that lets us use Nothing for the fields not present in the smaller record.

type RSubset = RecSubset Rec

type REquivalent rs ss is js = (RSubset rs ss is, RSubset ss rs js)

Two record types are equivalent when they are subtypes of each other.

type (∈) r rs = RElem r rs (RIndex r rs)

A shorthand for RElem which supplies its index.

type (⊆) rs ss = RSubset rs ss (RImage rs ss)

A shorthand for RSubset which supplies its image.

type (≅) rs ss = REquivalent rs ss (RImage rs ss) (RImage ss rs)

A shorthand for REquivalent which supplies its images.

type (<:) rs ss = rs ⊆ ss

A non-unicode equivalent of (⊆).

type (:~:) rs ss = rs ≅ ss

A non-unicode equivalent of (≅).