Data.Vinyl.XRec

Description

A variant of Rec whose values have eliminated common syntactic clutter due to Identity, Compose, and ElField type constructors.

A common pain point with using Rec is the mandatory context of each value. A basic record might look like this, Identity "joe" :& Identity 23 :& RNil :: Rec Identity '[String, Int]. The Identity constructors are a nuisance, so we offer a way of avoiding them: "joe" ::& 23 ::& XRNil :: XRec Identity '[String,Int]. Facilities are provided for converting between XRec and Rec so that the Rec API is available even if you choose to use XRec for construction or pattern matching.

Documentation

type XRec f = Rec (XData f)

pattern (::&) :: HKD f r -> XRec f rs -> XRec f (r ': rs) infixr 7

pattern XRNil :: XRec f '[]

rmapX :: forall f g rs. (XRMap f g rs, IsoXRec f rs, IsoXRec g rs) => (forall a. HKD f a -> HKD g a) -> Rec f rs -> Rec g rs

Like rmap, but the supplied function is written against the HKD-simplified types. This is xrmap sandwiched in between fromXRec and toXRec.

rmapXEndo :: forall f rs. (XRMap f f rs, IsoXRec f rs) => (forall a. HKD f a -> HKD f a) -> Rec f rs -> Rec f rs

This is rmapX specialized to a type at which it does not change interpretation functor. This can help with type inference.

xrmap :: forall f g rs. XRMap f g rs => (forall a. HKD f a -> HKD g a) -> XRec f rs -> XRec g rs

This is rmap for XRec. We apply a natural transformation between interpretation functors to transport a record value between interpretations.

newtype XData t a

A wrapper for an HKD-simplified value. That is, noisy value constructors like Identity and Compose are ellided. This is used in the xrmapAux type class method, but may be ignored by users whose needs are met by xrmap and rmapX.

Constructors

XData

Fields

unX :: HKD t a

class XRMap f g rs where

The implementation of xrmap is broken into a type class to permit unrolling of the recursion across a record. The function mapped across the vector hides the HKD type family under a newtype constructor to help the type checker.

Methods

xrmapAux :: (forall a. XData f a -> XData g a) -> XRec f rs -> XRec g rs

class XRApply f g rs where

Like rapply: record of components f r -> g r may be applied to a record of f to get a record of g.

Methods

xrapply :: XRec (Lift (->) f g) rs -> XRec f rs -> XRec g rs

class IsoXRec f ts where

Conversion between XRec and Rec. It is convenient to build and consume XRec values to reduce syntactic noise, but Rec has a richer API that is difficult to build around the HKD type family.

Methods

fromXRec :: XRec f ts -> Rec f ts

toXRec :: Rec f ts -> XRec f ts

class IsoHKD f a where

Isomorphism between a syntactically noisy value and a concise one. For types like, Identity, we prefer to work with values of the underlying type without writing out the Identity constructor. For Compose f g a, aka (f :. g) a, we prefer to work directly with values of type f (g a).

This involves the so-called higher-kinded data type family. See http://reasonablypolymorphic.com/blog/higher-kinded-data for more discussion.

Minimal complete definition

Nothing

Associated Types

type HKD f a

type HKD f a = f a

Methods

unHKD :: HKD f a -> f a

default unHKD :: HKD f a ~ f a => HKD f a -> f a

toHKD :: f a -> HKD f a

default toHKD :: HKD f a ~ f a => f a -> HKD f a

rgetX :: forall a record f rs. (RecElem record a a rs rs (RIndex a rs), RecElemFCtx record f, IsoHKD f a) => record f rs -> HKD f a

Record field getter that pipes the field value through HKD to eliminate redundant newtype wrappings. Usage will typically involve a visible type application to the field type. The definition is similar to, getHKD = toHKD . rget.